Rename to ConventionalExtensions, the other one was not acceptable
While it got me started and made me chuckle because I thought of
It's Always Sunny — I then looked the name up and it seemed a lot
harsher than I'd thought it'd be.

So let's let go of the joke and settle for a clearer name that more
people can be in on too.