Maintain one loader for the `load_extensions` session
Fixes #1

If we set a loader on the main entrance to `load_extensions`, any
`load_extension`/`load_extensions` calls from within an extension can
be tracked in a Set so we avoid double-loading.

Essentially, if load_extensions tries to load :first and :second, but
:first includes `load_extension :second` we should omit loading :second
when the original `load_extensions` gets to :second in its
`each { load_one _1 }` loop.

This means we can avoid using `::Kernel.require` and `$LOADED_FEATURES`,
and it should give us better Zeitwerk interop, though I haven't tested
this yet.

Then once we exit the top-level `load_extensions` we can discard the
loader and all its state. Yay GC!